requirements.txt
pytorch-lightning
pyrender
opencv-python
yacs
scikit-image
einops
timm
dill
pandas
rich
hydra-core
hydra-submitit-launcher
hydra-colorlog
pyrootutils
webdataset
chump
networkx==3.2.1
roma
joblib
seaborn
wandb
appdirs
appnope
ffmpeg
cython
jsonlines
pytest
xtcocotools
loguru
optree
fvcore
black
pycocotools
tensorboard
huggingface_hub
numpy
matplotlib
